SUREFIRE-1031 try to delete the file as early as possible, use deleteOnExit only as a fallback

+++ maven-surefire-common/src/main/java/org/apache/maven/plugin/surefire/report/Utf8RecodingDeferredFileOutputStream.java
@@ -93,7 +93,19 @@ class Utf8RecodingDeferredFileOutputStream
     {
         if ( null != deferredFileOutputStream && null != deferredFileOutputStream.getFile() )
         {
-            deferredFileOutputStream.getFile().deleteOnExit();
+            try
+            {
+                deferredFileOutputStream.close();
+                if ( !deferredFileOutputStream.getFile().delete() )
+                {
+                    deferredFileOutputStream.getFile().deleteOnExit();
+                }
+            }
+            catch ( IOException ioe )
+            {
+                deferredFileOutputStream.getFile().deleteOnExit();
+
+            }
         }
     }
 }
